+++ examples/src/nrf52_dk.rs
@@ -282,7 +282,7 @@ mod duration {
 
     pub fn get_generic_integer() {
         let generic = duration::Generic::new(246_u32, Fraction::new(1, 2));
-        assert_eq!(generic.integer(), &246_u32);
+        assert_eq!(generic.integer(), 246_u32);
     }
 
     pub fn remainder() {
@@ -417,7 +417,7 @@ mod rate {
 
     pub fn get_generic_integer() {
         let generic = rate::Generic::new(246_u32, Fraction::new(1, 2));
-        assert_eq!(generic.integer(), &246_u32);
+        assert_eq!(generic.integer(), 246_u32);
     }
 
     pub fn remainder() {