+++ src/instant.rs
@@ -414,7 +414,7 @@ impl<Clock: crate::Clock> ops::Sub<Instant<Clock>> for Instant<Clock> {
     /// # fn try_now(&self) -> Result<Instant<Self>, embedded_time::clock::Error> {unimplemented!()}
     /// }
     ///
-    /// assert_eq!(*(Instant::<Clock>::new(5_001) - Instant::<Clock>::new(5_000)).integer(), 1);
+    /// assert_eq!((Instant::<Clock>::new(5_001) - Instant::<Clock>::new(5_000)).integer(), 1);
     /// ```
     ///
     /// # Panics